backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
supabase
sentence-transformers
psycopg[binary]
networkx
